pyproject.toml
[project]
name = "meeting-protocol-generator"
version = "0.3.0"
description = "Автоматизированный конвейер для преобразования аудиозаписей совещаний в структурированные протоколы"
authors = [
    {name = "Альтэон Шульц"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.110.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.3",
    "jsonschema>=4.23.0",
    "openai>=1.79.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.0",
    "pydub>=0.25.1",
    "pyyaml>=6.0.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.12.0",
    "python-multipart>=0.0.9",
    "replicate>=1.0.6",
    "requests>=2.32.3",
    "rich>=13.7.0",
    "socketio>=0.2.1",
    "tiktoken>=0.7.0",
    "typer>=0.9.0",
    "typing-extensions>=4.9.0",
    "uvicorn>=0.27.1",
    "websockets>=12.0",
    "werkzeug>=3.1.3",
]

[project.urls]
"Homepage" = "https://github.com/kostas-dev/meeting-ai-flow"

[tool.setuptools]
packages = ["app"]

[project.scripts]
meeting-protocol-generator = "app.cli:main"

[tool.pytest.ini_options]
norecursedirs = ["_archive", ".venv", ".*", "node_modules", "dist", "build", "venv"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
select = ["E", "F", "W", "I", "N", "B", "COM", "C4", "UP", "PT", "RET", "SIM", "ARG", "PL"]
ignore = ["E501", "B019", "COM812", "PLR2004", "PLR0913", "PLR0912", "PLR0911", "PLW0603"]
line-length = 100
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"tests/*" = ["E501", "F403", "F405", "N802"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
